skyblockgame/requirements.txt
coverage==7.5.0
flake8==7.0.0
pygame-ce==2.4.1
pylint==3.1.0
PyTMX==3.32
